+++ docs/development/changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Renamed scan-block operators to `Stream*` and made the module public (`streaming.py`) (#173)
 - Avoid unnecessary array copies in observation readers (#175)
+- Migrate multi-observation mapmaker to furax CG and support verbose mode (#176)
 
 ## [0.11.3] - 2026-07-14